requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
dateparser==1.1.1
fonttools==4.37.3
frozenlist==1.3.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.0
mplfinance==0.12.9b1
multidict==6.0.2
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pycodestyle==2.9.1
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
toml==0.10.2
tzdata==2022.2
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websockets==10.3
yarl==1.8.1
